fix(jssdk): typing
