+++ frontend/packages/sdk/src/utils/aes-256-gcm.ts
@@ -1,6 +1,5 @@
 import { Buffer } from 'buffer'
 import { hexAddPrefix, hexStripPrefix, hexToU8a } from '@polkadot/util'
-// @ts-ignore
 import { createCipheriv, createDecipheriv } from 'browserify-cipher'
 
 const ALGO = 'aes-256-gcm'
